refactor: untangle bundled channel sdk bridges
