+++ extensions/discord/src/account-inspect.ts
@@ -1,13 +1,13 @@
 import {
+  hasConfiguredSecretInput,
+  normalizeSecretInputString,
+} from "../../../src/config/types.secrets.js";
+import {
   DEFAULT_ACCOUNT_ID,
   normalizeAccountId,
   type OpenClawConfig,
   type DiscordAccountConfig,
-} from "openclaw/plugin-sdk/discord";
-import {
-  hasConfiguredSecretInput,
-  normalizeSecretInputString,
-} from "../../../src/config/types.secrets.js";
+} from "../../../src/plugin-sdk-internal/discord.js";
 import {
   mergeDiscordAccountConfig,
   resolveDefaultDiscordAccountId,
