+++ extensions/discord/src/accounts.ts
@@ -1,14 +1,14 @@
-import type {
-  OpenClawConfig,
-  DiscordAccountConfig,
-  DiscordActionConfig,
-} from "openclaw/plugin-sdk/discord";
 import {
   createAccountActionGate,
   createAccountListHelpers,
   normalizeAccountId,
   resolveAccountEntry,
 } from "../../../src/plugin-sdk-internal/accounts.js";
+import type {
+  OpenClawConfig,
+  DiscordAccountConfig,
+  DiscordActionConfig,
+} from "../../../src/plugin-sdk-internal/discord.js";
 import { resolveDiscordToken } from "./token.js";
 
 export type ResolvedDiscordAccount = {
