+++ extensions/imessage/src/accounts.ts
@@ -1,10 +1,10 @@
-import type { IMessageAccountConfig } from "openclaw/plugin-sdk/imessage";
 import {
   type OpenClawConfig,
   createAccountListHelpers,
   normalizeAccountId,
   resolveAccountEntry,
 } from "../../../src/plugin-sdk-internal/accounts.js";
+import type { IMessageAccountConfig } from "../../../src/plugin-sdk-internal/imessage.js";
 
 export type ResolvedIMessageAccount = {
   accountId: string;
