+++ extensions/imessage/src/target-parsing-helpers.ts
@@ -1,4 +1,4 @@
-import { isAllowedParsedChatSender } from "../../../src/plugin-sdk/allow-from.js";
+import { isAllowedParsedChatSender } from "../../../src/plugin-sdk-internal/imessage.js";
 
 export type ServicePrefix<TService extends string> = { prefix: string; service: TService };
