+++ extensions/slack/src/message-action-dispatch.ts
@@ -1,1 +1,1 @@
-export { handleSlackMessageAction } from "../../../src/plugin-sdk/slack-message-actions.js";
+export { handleSlackMessageAction } from "../../../src/plugin-sdk-internal/slack.js";
