+++ extensions/telegram/src/token.ts
@@ -1,8 +1,8 @@
-import type { TelegramAccountConfig } from "openclaw/plugin-sdk/telegram";
 import type { BaseTokenResolution } from "../../../src/channels/plugins/types.core.js";
 import type { OpenClawConfig } from "../../../src/config/config.js";
 import { normalizeResolvedSecretInputString } from "../../../src/config/types.secrets.js";
 import { tryReadSecretFileSync } from "../../../src/infra/secret-file.js";
+import type { TelegramAccountConfig } from "../../../src/plugin-sdk-internal/telegram.js";
 import { DEFAULT_ACCOUNT_ID, normalizeAccountId } from "../../../src/routing/session-key.js";
 
 export type TelegramTokenSource = "env" | "tokenFile" | "config" | "none";
