+++ src/channels/plugins/contracts/suites.ts
@@ -402,7 +402,6 @@ export function installChannelDirectoryContractSuite(params: {
     if (params.invokeLookups === false) {
       return;
     }
-
     const self = await directory?.self?.({
       cfg: {} as OpenClawConfig,
       accountId: "default",
