+++ src/plugin-sdk-internal/accounts.ts
@@ -3,6 +3,10 @@ export type { OpenClawConfig } from "../config/config.js";
 export { createAccountActionGate } from "../channels/plugins/account-action-gate.js";
 export { createAccountListHelpers } from "../channels/plugins/account-helpers.js";
 export { normalizeChatType } from "../channels/chat-type.js";
+export {
+  listConfiguredAccountIds,
+  resolveAccountWithDefaultFallback,
+} from "../plugin-sdk/account-resolution.js";
 export { resolveAccountEntry } from "../routing/account-lookup.js";
 export { DEFAULT_ACCOUNT_ID, normalizeAccountId } from "../routing/session-key.js";
 export { normalizeE164, pathExists, resolveUserPath } from "../utils.js";
