+++ src/plugin-sdk-internal/channel-config.ts
@@ -1,0 +1,17 @@
+// Private bridge for bundled channel plugins. These config helpers are shared
+// internally, but do not belong on the public compat surface.
+export { buildAccountScopedAllowlistConfigEditor } from "../plugin-sdk/allowlist-config-edit.js";
+export { formatAllowFromLowercase } from "../plugin-sdk/allow-from.js";
+export {
+  createScopedAccountConfigAccessors,
+  createScopedChannelConfigBase,
+  createScopedDmSecurityResolver,
+} from "../plugin-sdk/channel-config-helpers.js";
+export {
+  collectAllowlistProviderGroupPolicyWarnings,
+  collectAllowlistProviderRestrictSendersWarnings,
+  collectOpenGroupPolicyConfiguredRouteWarnings,
+  collectOpenGroupPolicyRouteAllowlistWarnings,
+  collectOpenProviderGroupPolicyWarnings,
+} from "../channels/plugins/group-policy-warnings.js";
+export { buildAccountScopedDmSecurityPolicy } from "../channels/plugins/helpers.js";
